requirements.txt
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
isort==5.5.2
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.2
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
six==1.15.0
toml==0.10.1
twilio==6.53.0
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
